Refactor  - Break large singleton file into multiple files
