+++ ninchatsdk/src/main/java/com/ninchat/sdk/activities/NinchatActivity.java
@@ -10,27 +10,16 @@ import android.support.annotation.Nullable;
 import android.support.v4.content.LocalBroadcastManager;
 import android.support.v7.widget.RecyclerView;
 import android.text.TextUtils;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
-import com.ninchat.client.JSON;
-import com.ninchat.client.Props;
 import com.ninchat.sdk.NinchatSession;
 import com.ninchat.sdk.NinchatSessionManager;
 import com.ninchat.sdk.R;
 import com.ninchat.sdk.adapters.NinchatQueueListAdapter;
-import com.ninchat.sdk.tasks.NinchatRegisterAudienceTask;
-
-import org.json.JSONException;
-import org.json.JSONObject;
-
 import static com.ninchat.sdk.activities.NinchatQuestionnaireActivity.OPEN_QUEUE;
-import static com.ninchat.sdk.activities.NinchatQuestionnaireActivity.QUESTIONNAIRE_TYPE;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.POST_AUDIENCE_QUESTIONNAIRE;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.PRE_AUDIENCE_QUESTIONNAIRE;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getPreAnswers;
+import static com.ninchat.sdk.helper.questionnaire.NinchatQuestionnaireTypeUtil.*;
 
 public final class NinchatActivity extends NinchatBaseActivity {
