+++ ninchatsdk/src/main/java/com/ninchat/sdk/adapters/NinchatQueueListAdapter.java
@@ -17,7 +17,7 @@ import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.PRE_AUDIENCE_QUESTIONNAIRE;
+import static com.ninchat.sdk.helper.questionnaire.NinchatQuestionnaireTypeUtil.*;
 
 public final class NinchatQueueListAdapter extends RecyclerView.Adapter<NinchatQueueViewHolder> {
