+++ ninchatsdk/src/main/java/com/ninchat/sdk/adapters/holders/formview/NinchatInputFieldViewHolder.java
@@ -17,12 +17,9 @@ import org.json.JSONObject;
 
 import java.lang.ref.WeakReference;
 
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getError;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getLabel;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getResultString;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.matchPattern;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.setError;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.setResult;
+import static com.ninchat.sdk.helper.questionnaire.NinchatQuestionnaireItemGetter.*;
+import static com.ninchat.sdk.helper.questionnaire.NinchatQuestionnaireItemSetter.*;
+import static com.ninchat.sdk.helper.questionnaire.NinchatQuestionnaireMiscUtil.matchPattern;
 
 public class NinchatInputFieldViewHolder extends RecyclerView.ViewHolder {
     private final String TAG = NinchatInputFieldViewHolder.class.getSimpleName();
