+++ ninchatsdk/src/main/java/com/ninchat/sdk/models/questionnaire/conversation/NinchatConversationQuestionnaire.java
@@ -28,28 +28,10 @@ import org.json.JSONObject;
 import java.lang.ref.WeakReference;
 import java.util.Stack;
 
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.POST_AUDIENCE_QUESTIONNAIRE;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.clearElement;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.clearElementResult;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.formHasError;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getAllFilledElements;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getCurrentElement;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getFinalAnswers;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getMatchingLogic;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getMatchingLogicTarget;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getNextElementIndex;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getPreAnswers;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getQuestionnaire;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getQuestionnaireAnswers;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getQuestionnaireAnswersQueue;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getQuestionnaireAnswersTags;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.getQuestionnaireElementIndexByName;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.isComplete;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.isRegister;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.setQueue;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.setTags;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.setViewAndChildrenEnabled;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.updateRequiredFieldStats;
+import static com.ninchat.sdk.helper.questionnaire.NinchatQuestionnaireItemGetter.*;
+import static com.ninchat.sdk.helper.questionnaire.NinchatQuestionnaireItemSetter.*;
+import static com.ninchat.sdk.helper.questionnaire.NinchatQuestionnaireNavigationUtil.*;
+import static com.ninchat.sdk.helper.questionnaire.NinchatQuestionnaireTypeUtil.*;
 
 public class NinchatConversationQuestionnaire {
 
@@ -110,7 +92,7 @@ public class NinchatConversationQuestionnaire {
     private void close(final boolean isRegister) {
         final JSONObject answerList = getQuestionnaireAnswers(mQuestionnaire.getQuestionnaireList(), historyList);
         final JSONArray tagList = getQuestionnaireAnswersTags(mQuestionnaire.getQuestionnaireList(), historyList);
-        final JSONObject answers = getFinalAnswers(answerList, tagList);
+        final JSONObject answers = mergeAnswersAndTags(answerList, tagList);
         boolean openQueueView = true;
         if (questionnaireType == POST_AUDIENCE_QUESTIONNAIRE) {
             // a post audience questionnaire
